Allow updating a domain's associated project
